adding article body content, and adding in pages directory for Homepage
